improve Passenger / sub URI documentation
Talk a little more explicitly about sub URI configuration as it took me a while
to dig through the Passenger documentation and find the right buttons to push.
Be clearer about user permissions as the internet was full of terrible advice.
Suggest configuration is stored in /var/local which is reserved for non-package
managed software.